Create README.md
Sentiment analysis of Donald Trump Tweets.
Does Donald Trump's use of Twitter keep him popular or influence followers?

This article requires a classification based on tweets previously posted by former US President Donald Trump, which is used to perform sentiment analysis on the texts generated on the social network Twitter. The study aims to conduct sentiment analysis on tweets published in news outlets to understand how positive and negative Trump's tweets influence followers. For the development of sentiment analysis, the best way found was to use python libraries such as Vader Sentiment to extract the positive, negative, and neutral sentiments from Trump's tweets. NLTK import tokenize, lemmatize to develop algorithms that transform the text into a list of words and show the adjectives in the text. In addition, I used libraries like Matplotlib, WordCloud to generate charts and created charts in Tableau to show the adjectives Donald Trump used most in tweets, most frequent word cloud, correlation of feelings with the number of retweets and favourites and to observe the popularity of these tweets. All this following the steps of lexicon techniques, such as POS tagging, Stop words, etc.